+++ Main/Main/cat-service/src/main/resources/application.properties
@@ -1,5 +1,5 @@
-spring.application.name=lab3
-server.port=8080
+spring.application.name=cat-service
+server.port=8087
 
 spring.thymeleaf.cache=false
 spring.thymeleaf.enabled=true
@@ -9,8 +9,18 @@ spring.thymeleaf.suffix=.html
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5431/postgres
-spring.datasource.username=postgres
-spring.datasource.password=postgres
+spring.datasource.url=jdbc:postgresql://localhost:5432/main_db
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+spring.datasource.username=db_user
+spring.datasource.password=db_pass
 
 logging.level.org.springframework.security=TRACE
+
+spring.kafka.bootstrap-servers = localhost:9092
+
+spring.kafka.catCreateTopic=cats-create
+spring.kafka.catFindAllTopic=cats-find-all
+spring.kafka.catUpdateTopic=cats-update
+spring.kafka.catDeleteTopic=cats-delete
+
+spring.kafka.consumer.group-id = 3
