+++ Main/Main/cat-service/src/main/resources/application.yml
@@ -1,0 +1,29 @@
+server:
+  port: 8087
+
+spring:
+  application:
+    name: cat-service
+
+  kafka:
+    bootstrap-servers: localhost:9092
+    consumer:
+      group-id: 3
+      auto-offset-reset: earliest
+      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
+      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
+      properties:
+        spring.json.trusted.packages: "Models.CatDTO"
+    producer:
+      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
+
+  datasource:
+    url: jdbc:postgresql://localhost:5432/main_db
+    username: db_user
+    password: db_pass
+    driver-class-name: org.postgresql.Driver
+  jpa:
+    hibernate:
+      ddl-auto: update
+    show-sql: true
