+++ Main/Main/user-service/src/main/java/Main/Entities/Privilege.java
@@ -1,0 +1,27 @@
+package Main.Entities;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+
+import java.util.Collection;
+
+@Entity
+@Data
+@RequiredArgsConstructor
+public class Privilege {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    public Privilege(String name) {
+        this.name = name;
+    }
+
+    @Column(name = "name")
+    private String name;
+
+    @ManyToMany(mappedBy = "privileges")
+    private Collection<Role> roles;
+}
