+++ Main/Main/user-service/src/main/java/Main/Repository/PrivilegeRepository.java
@@ -1,0 +1,9 @@
+package Main.Repository;
+
+
+import Main.Entities.Privilege;
+import org.springframework.data.repository.CrudRepository;
+
+public interface PrivilegeRepository extends CrudRepository<Privilege, Long> {
+    Privilege findByName(String name);
+}
