+++ Main/Main/user-service/src/main/java/Main/Repository/RoleRepository.java
@@ -1,0 +1,11 @@
+package Main.Repository;
+
+
+import Main.Entities.Role;
+import org.springframework.data.repository.CrudRepository;
+
+public interface RoleRepository extends CrudRepository<Role, Long> {
+
+
+    Role findByName(String roleAdmin);
+}
