+++ Main/Main/user-service/src/main/java/Main/Repository/UserRepository.java
@@ -1,0 +1,16 @@
+package Main.Repository;
+
+
+import Main.Entities.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Long> {
+    User findByName(String name);
+
+    List<User> findAll();
+
+}
