--- Main/Main/user-service/src/main/java/Repository/PrivilegeRepository.java
@@ -1,9 +1,0 @@
-package Repository;
-
-
-import Entities.Privilege;
-import org.springframework.data.repository.CrudRepository;
-
-public interface PrivilegeRepository extends CrudRepository<Privilege, Long> {
-    Privilege findByName(String name);
-}
