--- Main/Main/user-service/src/main/java/Repository/RoleRepository.java
@@ -1,11 +1,0 @@
-package Repository;
-
-
-import Entities.Role;
-import org.springframework.data.repository.CrudRepository;
-
-public interface RoleRepository extends CrudRepository<Role, Long> {
-
-
-    Role findByName(String roleAdmin);
-}
