--- Main/Main/user-service/src/main/java/Repository/UserRepository.java
@@ -1,12 +1,0 @@
-package Repository;
-
-
-import Entities.User;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-    User findByName(String name);
-
-}
