--- Main/Main/user-service/src/main/java/Service/SecurityConfig.java
@@ -1,72 +1,0 @@
-package Service;
-
-import jakarta.servlet.DispatcherType;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.ProviderManager;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-public class SecurityConfig {
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-
-        http.csrf(AbstractHttpConfigurer::disable)
-                .authorizeHttpRequests(auth -> auth
-                        .dispatcherTypeMatchers(DispatcherType.ERROR).permitAll()
-                        .dispatcherTypeMatchers(DispatcherType.FORWARD).authenticated()
-                        .requestMatchers("/users/reg").permitAll()
-                        .requestMatchers("/users/login").permitAll()
-                        .requestMatchers(HttpMethod.GET, "/users/**").permitAll()
-                        .requestMatchers(HttpMethod.GET, "/cats/**").permitAll()
-                )
-                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                .httpBasic(Customizer.withDefaults());
-
-
-        http.authorizeHttpRequests(auth -> auth
-                .requestMatchers(HttpMethod.DELETE, "/cats").hasRole("ADMIN")
-                .requestMatchers("/cats").authenticated()
-                .requestMatchers("/cats/**").authenticated()
-
-        );
-
-        http.authorizeHttpRequests(auth -> auth
-                .requestMatchers(HttpMethod.DELETE, "/users").hasRole("ADMIN")
-                .requestMatchers("/users").authenticated()
-                .requestMatchers("/users/**").authenticated()
-
-        );
-
-        http.authorizeHttpRequests(auth -> auth.anyRequest().denyAll());
-
-        return http.build();
-    }
-
-
-    @Bean
-    public AuthenticationManager authenticationManager(
-            UserDetailsService userDetailsService,
-            PasswordEncoder passwordEncoder
-    ) {
-        DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
-        provider.setUserDetailsService(userDetailsService);
-        provider.setPasswordEncoder(passwordEncoder);
-        return new ProviderManager(provider);
-    }
-}
