+++ Main/Main/user-service/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.application.name=user-service
-server.port=8080
+server.port=8086
 
 spring.thymeleaf.cache=false
 spring.thymeleaf.enabled=true
@@ -9,17 +9,27 @@ spring.thymeleaf.suffix=.html
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5431/postgres
-spring.datasource.username=postgres
-spring.datasource.password=postgres
+spring.datasource.url=jdbc:postgresql://localhost:5432/main_db
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+spring.datasource.username=db_user
+spring.datasource.password=db_pass
 
-logging.level.org.springframework.security=TRACE
+#logging.level.org.springframework.security=TRACE
 
-spring.kafka.bootstrap-servers = localhost:/9092
+spring.kafka.bootstrap-servers = localhost:9092
 
 spring.kafka.userCreateTopic=users-create
 spring.kafka.userFindAllTopic=users-find-all
 spring.kafka.userUpdateTopic=users-update
 spring.kafka.userDeleteTopic=users-delete
 
-spring.kafka.consumer.group-id = 1
+spring.kafka.consumer.group-id = 2
+
+spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
+spring.kafka.producer.value-serializer=org.springframework.kafka.support.serializer.JsonSerializer
+
+
+spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+spring.kafka.consumer.value-deserializer=org.springframework.kafka.support.serializer.JsonDeserializer
+spring.kafka.consumer.properties.spring.json.trusted.packages=*
+spring.kafka.consumer.properties.spring.json.use.type.headers=false
