--- Main/Main/user-service/src/main/resources/application.yaml
@@ -1,34 +1,0 @@
-server:
-  port: 8081
-
-spring:
-  application:
-    name: owner-service
-
-  kafka:
-    bootstrap-servers: localhost:9092
-    consumer:
-      group-id: owner-service-group
-      auto-offset-reset: earliest
-      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
-      properties:
-        spring.json.trusted.packages: "com.example.owner.dto"
-    producer:
-      key-serializer: org.apache.kafka.common.serialization.StringSerializer
-      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
-
-  datasource:
-    url: jdbc:postgresql://localhost:5432/owner_db
-    username: postgres
-    password: postgres
-    driver-class-name: org.postgresql.Driver
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
-
-kafka:
-  topics:
-    owner-requests: owners.request
-    owner-replies: owners.reply
