+++ Main/Main/web-gateway/src/main/resources/application.properties
@@ -1,29 +1,65 @@
-spring.application.name=web-gateway
-server.port=8080
+server.port=8085
+server.servlet.context-path=/api
+
+spring.application.name=web-gateway-service
+
+# Kafka configuration
+spring.kafka.bootstrap-servers=localhost:9092
+spring.kafka.properties.security.protocol=PLAINTEXT
+
+# Producer settings
+spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
+spring.kafka.producer.value-serializer=org.springframework.kafka.support.serializer.JsonSerializer
+spring.kafka.producer.properties.acks=1
+spring.kafka.producer.properties.spring.json.add.type.headers=true
+
+# Consumer settings
+spring.kafka.consumer.group-id=1
+spring.kafka.consumer.auto-offset-reset=earliest
+spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+spring.kafka.consumer.value-deserializer=org.springframework.kafka.support.serializer.JsonDeserializer
+
+spring.kafka.consumer.properties.spring.json.trusted.packages=*
+spring.kafka.consumer.properties.spring.json.use.type.headers=true
+
+# Database configuration
+#spring.datasource.url=jdbc:postgresql://localhost:5432/main_db
+#spring.datasource.username=db_user
+#spring.datasource.password=db_pass
+#spring.datasource.driver-class-name=org.postgresql.Driver
+
+# JPA settings
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+
+# Security (OAuth2)
+spring.security.oauth2.resourceserver.jwt.issuer-uri=http://keycloak:8080/realms/myrealm
+spring.security.oauth2.resourceserver.jwt.jwk-set-uri=http://keycloak:8080/realms/myrealm/protocol/openid-connect/certs
+
+# Logging configuration
+logging.level.root=INFO
+logging.level.org.springframework.web=DEBUG
+logging.level.org.apache.kafka=WARN
+logging.level.com.example.gateway=DEBUG
+logging.file.name=logs/gateway.log
+
 
 spring.thymeleaf.cache=false
 spring.thymeleaf.enabled=true
 spring.thymeleaf.prefix=classpath:/templates/
 spring.thymeleaf.suffix=.html
 
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5431/postgres
-spring.datasource.username=postgres
-spring.datasource.password=postgres
 
 logging.level.org.springframework.security=TRACE
 
-spring.kafka.bootstrap-servers = localhost:/9092
 
-spring.kafka.consumer.group-id = 1
-
-
+#topics
 spring.kafka.catCreateTopic=cats-create
 spring.kafka.catFindAllTopic=cats-find-all
 spring.kafka.catUpdateTopic=cats-update
 spring.kafka.catDeleteTopic=cats-delete
+
+spring.kafka.userRepliesTopic=users-replies
 
 spring.kafka.userCreateTopic=users-create
 spring.kafka.userFindAllTopic=users-find-all
@@ -32,4 +68,4 @@ spring.kafka.userDeleteTopic=users-delete
 
 spring.kafka.globalReplicationFactor = 1
 spring.kafka.globalPartition = 1
-sptring.kafka.timeout=5
+spring.kafka.timeout=5
